// src/types/koneApi.ts
export type AccessToken = string
/**
 * ISO timestamp string in format '2021-03-22T13:52:57.412Z'
 */
export type Timestamp = string
/**
 * Globally unique identifier for the building in the format `building:123456`
 */
export type BuildingId = string
/**
 * Globally unique identifier for the lift group in the format `group:123456:1`
 */
export type GroupId = string
// ...
export type AreaId = string
/**
 * Globally unique identifier for the door in the format `door:123456:1:1:1:1`
 */
export type DoorId = string
/**
 * Call type of a lift call
 */
export type CallType = 'normal' | 'robot' | 'action'
/**
 * Call action of a lift call
 */
export type CallAction = 'destination'
// ...
export type RequestType = 'lift-call' | 'create-session' | 'resume-session' | 'lift-call-api-v2'

/**
 * Possible monitor events when making a lift call
 */
export type MonitorEvents = Array<'call' | 'deck' | 'door'>

/**
 * Unique request Identifier
 */
export type RequestId = string

/**
 * Unique session Identifier
 */
export type SessionId = string

/**
 * The time in seconds until the latest state data is re-sent
 */
export type ResendLatestStateUpToSeconds = number

/**
 *  amount of time it takes for the passenger to arrive to an elevator. Accepted range is from 1 to 600 seconds.
 */
export type PassengerArrivalTimeSeconds = number

/**
 *  Response type received by making a lift call, creating a session, resuming a session
 */
export type ResponseType = 'ok' | 'error'

/**
 *  Connection identifier
 */
export type ConnectionId = string

/**
 *  either close or keep connection open
 */
export type KeepAlive = boolean

export type CallPayload = {
  request_id: number
  area: number
  time: Timestamp
  terminal: number
  call: {
      action: number
      destination?: number
  }
}

/**
 * Destination call payload
 */
export type DestinationCallPayload = {
  type: RequestType
  callType: CallType
  callAction?: CallAction
  requestId?: RequestId
  buildingId: BuildingId
  groupId?: GroupId
  payload?: CallPayload
  sourceId?: AreaId
  destinationId?: AreaId
  monitorEvents?: MonitorEvents
  keepAlive?: KeepAlive
  passengerArrivalTimeSeconds?: PassengerArrivalTimeSeconds
}